Omit some boundary patterns of UUIDv4 reversibleness
